Relicense repo to MIT / CC BY 4.0
The code in this repository was previously licensed as GPLv3. Many
of the Jupyter notebooks contain snippets of code that can be useful
in all sorts of applications. The GPLv3 is counter productive, and
a permissive license that allows copying these snippets in any kind
of project is more suitable.

To the extent possible, the source code in the repository is now
licensed as MIT and the data as CC BY 4.0.